requirements.txt
pycryptodome==3.20.0
ecdsa==0.19.0